# docs/storage_swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "0G Storage Scan API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/accounts/{address}": {
            "get": {
                "description": "Query account information for specified account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Account's information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The account address",
                        "name": "address",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.AccountInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
        "/accounts/{address}/txs": {
            "get": {
                "description": "Query storage transactions for specified account, support root hash filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Account's storage transaction list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The submitter address of the uploaded file",
                        "name": "address",
                        "in": "path"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The merkle root hash of the uploaded file",
                        "name": "rootHash",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The layer1 tx hash of the submission",
                        "name": "txHash",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp in seconds",
                        "name": "minTimestamp",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Timestamp in seconds",
                        "name": "maxTimestamp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.StorageTxList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
        "/miners": {
            "get": {
                "description": "Query miners",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miner reward"
                ],
                "summary": "Miner list",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort in ASC or DESC order by timestamp(or relevant field)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "latest_update_time",
                            "total_reward"
                        ],
                        "type": "string",
                        "default": "latest_update_time",
                        "description": "The field used for sorting. The value is latest_update_time or total_reward",
                        "name": "sortField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.MinerList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
        "/miners/{address}": {
            "get": {
                "description": "Query miner information for specified account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miner reward"
                ],
                "summary": "Miner's information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The miner address",
                        "name": "address",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.MinerInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
        "/miners/{address}/rewards": {
            "get": {
                "description": "Query rewards for specified miner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miner reward"
                ],
                "summary": "Miner's reward list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The submitter address of the uploaded file",
                        "name": "address",
                        "in": "path"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.RewardList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
...
        "/stats/summary": {
            "get": {
                "description": "Query statistics summary includes `storage fee` and `log sync height`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistic"
                ],
                "summary": "Statistics summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "Data": {
                                            "$ref": "#/definitions/storage.Summary"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/api.BusinessError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.BusinessError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "storage.AccountInfo": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "The balance in layer 1",
                    "type": "number"
                },
                "dataSize": {
                    "description": "Total Size of storage data",
                    "type": "integer"
                },
                "expiredFiles": {
                    "description": "The number of expired files",
                    "type": "integer"
                },
                "files": {
                    "description": "Total number of files",
                    "type": "integer"
                },
                "prunedFiles": {
                    "description": "The number of pruned files",
                    "type": "integer"
                },
                "storageFee": {
                    "description": "Total storage fee",
                    "type": "number"
                },
                "totalReward": {
                    "description": "Total amount of distributed rewards",
                    "type": "number"
                },
                "txs": {
                    "description": "Total number of layer1 transaction",
                    "type": "integer"
                }
            }
        },
...
        "storage.StorageTxList": {
            "description": "Submission information list",
            "type": "object",
            "properties": {
                "list": {
                    "description": "Submission list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.StorageTxInfo"
                    }
                },
                "total": {
                    "description": "The total number of submission returned",
                    "type": "integer"
                }
            }
        },
...
    }
}